Ajustes salvar compra e tela de consulta da compra

+++ Modelos/ModeloCompra.cs
@@ -8,12 +8,12 @@ namespace Modelos
 {
     public class ModeloCompra
     {
-        public int CompraId { get; set; }
+        public Int64 CompraId { get; set; }
         public string DataCadastro { get; set; }
-        public int NumNota { get; set; }
+        public Int64 NumNota { get; set; }
         public DateTime DataNota { get; set; }
         public decimal PrecoNota { get; set; }
-        public int FornecedorId { get; set; }
+        public Int64 FornecedorId { get; set; }
         public string Observacao { get; set; }
         public char Status { get; set; }
     }
